Factor out map tooltip and context overlay components
